+++ pycantonese/pos_tagging/__init__.py
@@ -1,0 +1,4 @@
+from pycantonese.pos_tagging.hkcancor_to_ud import hkcancor_to_ud
+
+
+__all__ = ["hkcancor_to_ud"]